putting the current functions in the correct files
